FIx chat refresh + add stop button
