+++ web/src/app/chat/ChatPage.tsx
@@ -1,226 +1,47 @@
-import {
-  AuthTypeMetadata,
-  getAuthTypeMetadataSS,
-  getCurrentUserSS,
-} from "@/lib/userSS";
-import { redirect } from "next/navigation";
-import { fetchSS } from "@/lib/utilsSS";
-import { Connector, DocumentSet, User, ValidSources } from "@/lib/types";
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import { ChatSession } from "./interfaces";
 import { ChatSidebar } from "./sessionSidebar/ChatSidebar";
 import { Chat } from "./Chat";
-import {
-  BackendMessage,
-  ChatSession,
-  Message,
-  RetrievalType,
-} from "./interfaces";
-import { unstable_noStore as noStore } from "next/cache";
+import { DocumentSet, User, ValidSources } from "@/lib/types";
 import { Persona } from "../admin/personas/interfaces";
-import { InstantSSRAutoRefresh } from "@/components/SSRAutoRefresh";
-import { WelcomeModal } from "@/components/WelcomeModal";
-import { ApiKeyModal } from "@/components/openai/ApiKeyModal";
-import { cookies } from "next/headers";
-import { DOCUMENT_SIDEBAR_WIDTH_COOKIE_NAME } from "@/components/resizable/contants";
-import { personaComparator } from "../admin/personas/lib";
 
-export default async function ChatPage({
-  chatId,
-  shouldhideBeforeScroll,
+export function ChatLayout({
+  user,
+  chatSessions,
+  availableSources,
+  availableDocumentSets,
+  availablePersonas,
+  documentSidebarInitialWidth,
 }: {
-  chatId: string | null;
-  shouldhideBeforeScroll?: boolean;
+  user: User | null;
+  chatSessions: ChatSession[];
+  availableSources: ValidSources[];
+  availableDocumentSets: DocumentSet[];
+  availablePersonas: Persona[];
+  documentSidebarInitialWidth?: number;
 }) {
-  noStore();
-
-  const currentChatId = chatId ? parseInt(chatId) : null;
-
-  const tasks = [
-    getAuthTypeMetadataSS(),
-    getCurrentUserSS(),
-    fetchSS("/manage/connector"),
-    fetchSS("/manage/document-set"),
-    fetchSS("/persona?include_default=true"),
-    fetchSS("/chat/get-user-chat-sessions"),
-    chatId !== null
-      ? fetchSS(`/chat/get-chat-session/${chatId}`)
-      : (async () => null)(),
-  ];
-
-  // catch cases where the backend is completely unreachable here
-  // without try / catch, will just raise an exception and the page
-  // will not render
-  let results: (User | Response | AuthTypeMetadata | null)[] = [
-    null,
-    null,
-    null,
-    null,
-    null,
-    null,
-    null,
-  ];
-  try {
-    results = await Promise.all(tasks);
-  } catch (e) {
-    console.log(`Some fetch failed for the main search page - ${e}`);
-  }
-  const authTypeMetadata = results[0] as AuthTypeMetadata | null;
-  const user = results[1] as User | null;
-  const connectorsResponse = results[2] as Response | null;
-  const documentSetsResponse = results[3] as Response | null;
-  const personasResponse = results[4] as Response | null;
-  const chatSessionsResponse = results[5] as Response | null;
-  const chatSessionMessagesResponse = results[6] as Response | null;
-
-  const authDisabled = authTypeMetadata?.authType === "disabled";
-  if (!authDisabled && !user) {
-    return redirect("/auth/login");
-  }
-
-  if (user && !user.is_verified && authTypeMetadata?.requiresVerification) {
-    return redirect("/auth/waiting-on-verification");
-  }
-
-  let connectors: Connector<any>[] = [];
-  if (connectorsResponse?.ok) {
-    connectors = await connectorsResponse.json();
-  } else {
-    console.log(`Failed to fetch connectors - ${connectorsResponse?.status}`);
-  }
-  const availableSources: ValidSources[] = [];
-  connectors.forEach((connector) => {
-    if (!availableSources.includes(connector.source)) {
-      availableSources.push(connector.source);
-    }
-  });
-
-  let chatSessions: ChatSession[] = [];
-  if (chatSessionsResponse?.ok) {
-    chatSessions = (await chatSessionsResponse.json()).sessions;
-  } else {
-    console.log(
-      `Failed to fetch chat sessions - ${chatSessionsResponse?.text()}`
-    );
-  }
-  // Larger ID -> created later
-  chatSessions.sort((a, b) => (a.id > b.id ? -1 : 1));
-  const currentChatSession = chatSessions.find(
-    (chatSession) => chatSession.id === currentChatId
-  );
-
-  let documentSets: DocumentSet[] = [];
-  if (documentSetsResponse?.ok) {
-    documentSets = await documentSetsResponse.json();
-  } else {
-    console.log(
-      `Failed to fetch document sets - ${documentSetsResponse?.status}`
-    );
-  }
-
-  let personas: Persona[] = [];
-  if (personasResponse?.ok) {
-    personas = await personasResponse.json();
-  } else {
-    console.log(`Failed to fetch personas - ${personasResponse?.status}`);
-  }
-  // remove those marked as hidden by an admin
-  personas = personas.filter((persona) => persona.is_visible);
-  // sort them in priority order
-  personas.sort(personaComparator);
-
-  let messages: Message[] = [];
-  if (chatSessionMessagesResponse?.ok) {
-    const chatSessionDetailJson = await chatSessionMessagesResponse.json();
-    const rawMessages = chatSessionDetailJson.messages as BackendMessage[];
-    const messageMap: Map<number, BackendMessage> = new Map(
-      rawMessages.map((message) => [message.message_id, message])
-    );
-
-    const rootMessage = rawMessages.find(
-      (message) => message.parent_message === null
-    );
-
-    const finalMessageList: BackendMessage[] = [];
-    if (rootMessage) {
-      let currMessage: BackendMessage | null = rootMessage;
-      while (currMessage) {
-        finalMessageList.push(currMessage);
-        const childMessageNumber = currMessage.latest_child_message;
-        if (childMessageNumber && messageMap.has(childMessageNumber)) {
-          currMessage = messageMap.get(childMessageNumber) as BackendMessage;
-        } else {
-          currMessage = null;
-        }
-      }
-    }
-
-    messages = finalMessageList
-      .filter((messageInfo) => messageInfo.message_type !== "system")
-      .map((messageInfo) => {
-        const hasContextDocs =
-          (messageInfo?.context_docs?.top_documents || []).length > 0;
-        let retrievalType;
-        if (hasContextDocs) {
-          if (messageInfo.rephrased_query) {
-            retrievalType = RetrievalType.Search;
-          } else {
-            retrievalType = RetrievalType.SelectedDocs;
-          }
-        } else {
-          retrievalType = RetrievalType.None;
-        }
-
-        return {
-          messageId: messageInfo.message_id,
-          message: messageInfo.message,
-          type: messageInfo.message_type as "user" | "assistant",
-          // only include these fields if this is an assistant message so that
-          // this is identical to what is computed at streaming time
-          ...(messageInfo.message_type === "assistant"
-            ? {
-                retrievalType: retrievalType,
-                query: messageInfo.rephrased_query,
-                documents: messageInfo?.context_docs?.top_documents || [],
-                citations: messageInfo?.citations || {},
-              }
-            : {}),
-        };
-      });
-  } else {
-    console.log(
-      `Failed to fetch chat session messages - ${chatSessionMessagesResponse?.text()}`
-    );
-  }
-
-  const documentSidebarCookieInitialWidth = cookies().get(
-    DOCUMENT_SIDEBAR_WIDTH_COOKIE_NAME
-  );
-  const finalDocumentSidebarInitialWidth = documentSidebarCookieInitialWidth
-    ? parseInt(documentSidebarCookieInitialWidth.value)
-    : undefined;
+  const searchParams = useSearchParams();
+  const chatIdRaw = searchParams.get("chatId");
+  const chatId = chatIdRaw ? parseInt(chatIdRaw) : null;
 
   return (
     <>
-      <InstantSSRAutoRefresh />
-      <ApiKeyModal />
-
-      {connectors.length === 0 && connectorsResponse?.ok && <WelcomeModal />}
-
       <div className="flex relative bg-background text-default h-screen overflow-x-hidden">
         <ChatSidebar
           existingChats={chatSessions}
-          currentChatId={currentChatId}
+          currentChatId={chatId}
           user={user}
         />
 
         <Chat
-          existingChatSessionId={currentChatId}
-          existingChatSessionPersonaId={currentChatSession?.persona_id}
-          existingMessages={messages}
+          existingChatSessionId={chatId}
+          existingChatSessionPersonaId={0}
           availableSources={availableSources}
-          availableDocumentSets={documentSets}
-          availablePersonas={personas}
-          documentSidebarInitialWidth={finalDocumentSidebarInitialWidth}
-          shouldhideBeforeScroll={shouldhideBeforeScroll}
+          availableDocumentSets={availableDocumentSets}
+          availablePersonas={availablePersonas}
+          documentSidebarInitialWidth={documentSidebarInitialWidth}
         />
       </div>
     </>
